backend/requirements.txt
Django==4.2.24
djangorestframework==3.16.1
django-cors-headers==4.7.0
python-decouple==3.8
requests==2.32.5
dj-database-url==2.1.0
psycopg2-binary==2.9.9
